+++ htdocs/inc_appdata/html_modal_templates.php
@@ -13,20 +13,52 @@ function write_modal_tpls(){
     global $que_modal_tpl__list, $_html_modal_templates;
     $html = '';
     foreach($que_modal_tpl__list as $key){
-        $html .= $_html_modal_templates[$key];
+
+        if($key=='login'){
+
+            //$html .= str_replace('{{URL_GOOGLE_LOGIN}}', $_SESSION['url_google_login'], $_html_modal_templates[$key]);
+            $html .= str_replace('{{URL_GOOGLE_LOGIN}}', 'index.php?google_login=' . rawurlencode(prepareLocation()), $_html_modal_templates[$key]);
+        } else if($key=='logout'){
+            if (!empty($_SESSION['USER_GOOGLE_LOGIN'])) {
+                [$a, $b, $data] = $_SESSION['USER_GOOGLE_LOGIN'];
+            } else {
+                $data = ['image'=>'/images/avatars/person.png','email'=>'example-email@gmail.com','name'=>'Ola Nordmann'];
+            }
+            $html .= str_replace(
+                ['{{USER_IMAGE}}',  '{{USER_EMAIL}}',   '{{USER_NAME}}'], 
+                [   $data['image'],    $data['email'],     $data['name']], 
+                $_html_modal_templates[$key]
+            );
+        } else {
+            $html .= $_html_modal_templates[$key];
+        } 
+
     }
     return '<!-- write_modal_tpls -->' . LF . minify_html($html, 'modals') . LF;
 }
 
 
-$client = new Google_Client();
-$client->setClientId($google_client_id);
-$client->setClientSecret($google_client_secret);
-$client->setRedirectUri('https://kista-ai.steinhaug.no/login.php');
-$client->addScope("email");
-$client->addScope("profile");
-$google_url = $client->createAuthUrl();
-logfile($google_url);
+if($lang == 'nb'){
+    $_l = [
+        'sign_in' => 'Logg inn',
+        'login_google' => 'Logg inn med Google',
+        'login_skip' => 'Fortsett uten innlogging',
+        'sign_out' => 'Utlogging',
+        'logout' => 'Logg ut',
+        'or' => 'eller',
+    ];
+} else {
+    $_l = [
+        'sign_in' => 'Sign In',
+        'login_google' => 'Sign In with Google',
+        'login_skip' => 'Continue without logging in',
+        'sign_out' => 'Sign Out',
+        'logout' => 'Logout',
+        'or' => 'or',
+    ];
+}
+
+
 
 $_html_modal_templates = [];
 $_html_modal_templates['login'] = '
@@ -35,25 +67,23 @@ $_html_modal_templates['login'] = '
          data-menu-height="310" 
          data-menu-width="350">
         <div class="menu-title">
-            <h1 class="font-24">Sign In</h1>
+            <h1 class="font-24">' . $_l['sign_in'] . '</h1>
             <a href="#" class="close-menu"><i class="fa fa-times-circle"></i></a>
         </div>
         <div class="content mb-0 mt-2">
-            <a href="' . $google_url . '" class="btn btn-border btn-m btn-full mb-3 rounded-s text-uppercase font-700 border-gray-dark color-night-dark bg-theme">
+            <a href="{{URL_GOOGLE_LOGIN}}" class="btn btn-border btn-m btn-full mb-3 rounded-s text-uppercase font-700 border-gray-dark color-night-dark bg-theme">
                 <img class="login-btn-icon" src="/app/icons/google-logo.svg" alt="Google Logo">
-                <span>Sign In with Google</span>
+                <span>' . $_l['login_google'] . '</span>
             </a>
-            <div class="divider-spanp" role="separator"><span><p>OR</p></span></div>
+            <div class="divider-spanp" role="separator"><span><p>' . strtoupper($_l['or']) . '</p></span></div>
             <a href="#" class="btn close-menu btn-full btn-m color-blue-dark border-blue-light font-600 rounded-s">
-                Continue without logging in
+                ' . $_l['login_skip'] . '
             </a>
         </div>
     </div>
 ';
 
-$src = 'https://lh3.googleusercontent.com/a/ACg8ocJNC3s8hzeoUVz8J20KfBQCd3lUJ3373b51AvuVau6KJ18U=s96-c';
-$name = 'Kim Steinhaug';
-$email = 'steinhaug@gmail.com';
+
 
 $_html_modal_templates['logout'] = '
     <div id="modalMenu-logout" 
@@ -61,19 +91,19 @@ $_html_modal_templates['logout'] = '
          data-menu-height="250" 
          data-menu-width="350">
         <div class="menu-title">
-            <h1 class="font-24">Sign Out</h1>
+            <h1 class="font-24">' . $_l['sign_out'] . '</h1>
             <a href="#" class="close-menu"><i class="fa fa-times-circle"></i></a>
         </div>
         <div class="content mb-0 mt-2">
             <div class="d-flex">
-                <div><img src="' . $src . '" class="me-3 rounded-circle shadow-l" width="50"></div>
+                <div><img src="{{USER_IMAGE}}" class="me-3 rounded-circle shadow-l" width="50"></div>
                 <div>
-                    <h5 class="mt-1 mb-0">' . $name . '</h5>
-                    <p class="font-10 mt-n1 color-gray-dark">' . $email . '</p>
+                    <h5 class="mt-1 mb-0">{{USER_NAME}}</h5>
+                    <p class="font-10 mt-n1 color-gray-dark">{{USER_EMAIL}}</p>
                 </div>
             </div>
             <br>
-            <p><a href="index.php?action=logout" class="close-menu btn btn-full btn-m shadow-l rounded-s text-uppercase font-600 gradient-blue mt-n2">Logout from site</a></p>
+            <p><a href="index.php?action=logout" class="close-menu btn btn-full btn-m shadow-l rounded-s text-uppercase font-600 gradient-blue mt-n2">' . $_l['logout'] . '</a></p>
         </div>
     </div>
 ';